refactor: extract shared runtime code into bootstrap-shared.js
Move dlopen patching, child_process patching, and process.pkg setup
into a shared module used by both the traditional and SEA bootstraps.
This eliminates duplication and ensures both modes handle native
addons, subprocess spawning, and process.pkg identically.

- Replace copyFolderRecursiveSync with fs.cpSync (Node >= 22)
- Add REQUIRE_SHARED parameter to packer wrapper for traditional mode
- SEA bootstrap consumes shared module via esbuild bundling
- Remove dead code (ARGV0, homedir import, copyFolderRecursiveSync)
