+++ prelude/bootstrap.js
@@ -2,6 +2,7 @@
 /* global PAYLOAD_POSITION */
 /* global PAYLOAD_SIZE */
 /* global REQUIRE_COMMON */
+/* global REQUIRE_SHARED */
 /* global VIRTUAL_FILESYSTEM */
 /* global DEFAULT_ENTRYPOINT */
 /* global DICT */
@@ -18,7 +19,6 @@ const Module = require('module');
 const path = require('path');
 const { promisify } = require('util');
 const { Script } = require('vm');
-const { homedir } = require('os');
 const util = require('util');
 const {
   brotliDecompress,
@@ -50,9 +50,8 @@ const NODE_VERSION_MINOR = process.version.match(/^v\d+.(\d+)/)[1] | 0;
 // ENTRYPOINT //////////////////////////////////////////////////////
 // /////////////////////////////////////////////////////////////////
 
-// set ENTRYPOINT and ARGV0 here because
-// they can be altered during process run
-const ARGV0 = process.argv[0];
+// set ENTRYPOINT here because
+// it can be altered during process run
 const EXECPATH = process.execPath;
 let ENTRYPOINT = process.argv[1];
 
@@ -171,74 +170,6 @@ function copyInChunks(
 
   fs_.closeSync(sourceFile);
   fs_.closeSync(targetFile);
-}
-
-// TODO: replace this with fs.cpSync when we drop Node < 16
-function copyFolderRecursiveSync(source, target) {
-  // Build target folder
-  const targetFolder = path.join(target, path.basename(source));
-
-  // Check if target folder needs to be created or integrated
-  fs.mkdirSync(targetFolder, { recursive: true });
-
-  // Copy
-  if (fs.lstatSync(source).isDirectory()) {
-    const files = fs.readdirSync(source);
-
-    for (const file of files) {
-      // Build source name
-      const curSource = path.join(source, file);
-
-      // Call this function recursively as long as source is a directory
-      if (fs.lstatSync(curSource).isDirectory()) {
-        copyFolderRecursiveSync(curSource, targetFolder);
-      } else {
-        // Current source is a file, it must be available on the real filesystem
-        // instead of the virtual snapshot file system to load it by process.dlopen.
-        //
-        // Before we try to copy we do some checks.
-        // See https://github.com/vercel/pkg/issues/1589 for more details.
-
-        // Build target file name
-        const curTarget = path.join(targetFolder, path.basename(curSource));
-
-        if (fs.existsSync(curTarget)) {
-          // Target file already exists, read source and target file...
-          const curSourceContent = fs.readFileSync(curSource, {
-            encoding: 'binary',
-          });
-          const curTargetContent = fs.readFileSync(curTarget, {
-            encoding: 'binary',
-          });
-
-          // ...and calculate checksum from source and target file
-          const curSourceHash = createHash('sha256')
-            .update(curSourceContent)
-            .digest('hex');
-          const curTargetHash = createHash('sha256')
-            .update(curTargetContent)
-            .digest('hex');
-
-          // If checksums are equal then there is nothing to do here
-          // ==> target already exists and is up-to-date
-          if (curSourceHash === curTargetHash) {
-            continue;
-          }
-        }
-
-        // Target must be copied because it either does not exist or is outdated.
-        // Due to the possibility that mutliple instances of this app start simultaneously,
-        // the copy action might fail. Only one starting instance gets write access.
-        //
-        // We don't catch any error here because it does not make sense to go ahead and to
-        // try to load the file while another instance has not yet finished the copy action.
-        // If the app start fails then the user should try to start the app later again.
-        // Unfortunately, we cannot implement delayed retries ourselves because process.dlopen
-        // is a synchronous function, promises are not supported.
-        fs.copyFileSync(curSource, curTarget);
-      }
-    }
-  }
 }
 
 /*
@@ -557,24 +488,9 @@ function payloadFileSync(pointer) {
 // /////////////////////////////////////////////////////////////////
 
 (() => {
-  process.pkg = {};
+  REQUIRE_SHARED.setupProcessPkg(ENTRYPOINT);
   process.versions.pkg = '%VERSION%';
   process.pkg.mount = createMountpoint;
-  process.pkg.entrypoint = ENTRYPOINT;
-  process.pkg.defaultEntrypoint = DEFAULT_ENTRYPOINT;
-})();
-
-// /////////////////////////////////////////////////////////////////
-// PATH.RESOLVE REPLACEMENT ////////////////////////////////////////
-// /////////////////////////////////////////////////////////////////
-
-(() => {
-  process.pkg.path = {};
-  process.pkg.path.resolve = function resolve() {
-    const args = cloneArgs(arguments);
-    args.unshift(path.dirname(ENTRYPOINT));
-    return path.resolve.apply(path, args);
-  };
 })();
 
 // /////////////////////////////////////////////////////////////////
@@ -1976,134 +1892,8 @@ function payloadFileSync(pointer) {
 // /////////////////////////////////////////////////////////////////
 // PATCH CHILD_PROCESS /////////////////////////////////////////////
 // /////////////////////////////////////////////////////////////////
-(() => {
-  const ancestor = {
-    spawn: childProcess.spawn,
-    spawnSync: childProcess.spawnSync,
-    execFile: childProcess.execFile,
-    execFileSync: childProcess.execFileSync,
-    exec: childProcess.exec,
-    execSync: childProcess.execSync,
-  };
 
-  function setOptsEnv(args) {
-    let pos = args.length - 1;
-    if (typeof args[pos] === 'function') pos -= 1;
-    if (typeof args[pos] !== 'object' || Array.isArray(args[pos])) {
-      pos += 1;
-      args.splice(pos, 0, {});
-    }
-    const opts = args[pos];
-    if (!opts.env) opts.env = { ...process.env };
-    // see https://github.com/vercel/pkg/issues/897#issuecomment-1049370335
-    if (opts.env.PKG_EXECPATH !== undefined) return;
-    opts.env.PKG_EXECPATH = EXECPATH;
-  }
-
-  function startsWith2(args, index, name, impostor) {
-    const qsName = `"${name} `;
-    if (args[index].slice(0, qsName.length) === qsName) {
-      args[index] = `"${impostor} ${args[index].slice(qsName.length)}`;
-      return true;
-    }
-    const sName = `${name} `;
-    if (args[index].slice(0, sName.length) === sName) {
-      args[index] = `${impostor} ${args[index].slice(sName.length)}`;
-      return true;
-    }
-    if (args[index] === name) {
-      args[index] = impostor;
-      return true;
-    }
-    return false;
-  }
-
-  function startsWith(args, index, name) {
-    const qName = `"${name}"`;
-    const qEXECPATH = `"${EXECPATH}"`;
-    const jsName = JSON.stringify(name);
-    const jsEXECPATH = JSON.stringify(EXECPATH);
-    return (
-      startsWith2(args, index, name, EXECPATH) ||
-      startsWith2(args, index, qName, qEXECPATH) ||
-      startsWith2(args, index, jsName, jsEXECPATH)
-    );
-  }
-
-  function modifyLong(args, index) {
-    if (!args[index]) return;
-    return (
-      startsWith(args, index, 'node') ||
-      startsWith(args, index, ARGV0) ||
-      startsWith(args, index, ENTRYPOINT) ||
-      startsWith(args, index, EXECPATH)
-    );
-  }
-
-  function modifyShort(args) {
-    if (!args[0]) return;
-    if (!Array.isArray(args[1])) {
-      args.splice(1, 0, []);
-    }
-    if (
-      args[0] === 'node' ||
-      args[0] === ARGV0 ||
-      args[0] === ENTRYPOINT ||
-      args[0] === EXECPATH
-    ) {
-      args[0] = EXECPATH;
-    } else {
-      for (let i = 1; i < args[1].length; i += 1) {
-        const mbc = args[1][i - 1];
-        if (mbc === '-c' || mbc === '/c') {
-          modifyLong(args[1], i);
-        }
-      }
-    }
-  }
-
-  childProcess.spawn = function spawn() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyShort(args);
-    return ancestor.spawn.apply(childProcess, args);
-  };
-
-  childProcess.spawnSync = function spawnSync() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyShort(args);
-    return ancestor.spawnSync.apply(childProcess, args);
-  };
-
-  childProcess.execFile = function execFile() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyShort(args);
-    return ancestor.execFile.apply(childProcess, args);
-  };
-
-  childProcess.execFileSync = function execFileSync() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyShort(args);
-    return ancestor.execFileSync.apply(childProcess, args);
-  };
-
-  childProcess.exec = function exec() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyLong(args, 0);
-    return ancestor.exec.apply(childProcess, args);
-  };
-
-  childProcess.execSync = function execSync() {
-    const args = cloneArgs(arguments);
-    setOptsEnv(args);
-    modifyLong(args, 0);
-    return ancestor.execSync.apply(childProcess, args);
-  };
-})();
+REQUIRE_SHARED.patchChildProcess(ENTRYPOINT);
 
 // /////////////////////////////////////////////////////////////////
 // PROMISIFY ///////////////////////////////////////////////////////
@@ -2176,79 +1966,5 @@ function payloadFileSync(pointer) {
 // /////////////////////////////////////////////////////////////////
 // PATCH PROCESS ///////////////////////////////////////////////////
 // /////////////////////////////////////////////////////////////////
-(() => {
-  const ancestor = {
-    dlopen: process.dlopen,
-  };
 
-  // Allow users to override the cache base directory via PKG_NATIVE_CACHE_PATH environment variable
-  // Default: path.join(homedir(), '.cache')
-  //   - Linux/macOS: /home/john/.cache or /Users/john/.cache
-  //   - Windows: C:\Users\John\.cache
-  // Custom example: /opt/myapp/cache or C:\myapp\cache
-  // Native addons will be extracted to: <PKG_NATIVE_CACHE_BASE>/pkg/<hash>
-  const PKG_NATIVE_CACHE_BASE =
-    process.env.PKG_NATIVE_CACHE_PATH || path.join(homedir(), '.cache');
-
-  function revertMakingLong(f) {
-    if (/^\\\\\?\\/.test(f)) return f.slice(4);
-    return f;
-  }
-
-  process.dlopen = function dlopen() {
-    const args = cloneArgs(arguments);
-    const modulePath = revertMakingLong(args[1]);
-    const moduleBaseName = path.basename(modulePath);
-    const moduleFolder = path.dirname(modulePath);
-
-    if (insideSnapshot(modulePath)) {
-      const moduleContent = fs.readFileSync(modulePath);
-
-      // Node addon files and .so cannot be read with fs directly, they are loaded with process.dlopen which needs a filesystem path
-      // we need to write the file somewhere on disk first and then load it
-      // the hash is needed to be sure we reload the module in case it changes
-      const hash = createHash('sha256').update(moduleContent).digest('hex');
-
-      const tmpFolder = path.join(PKG_NATIVE_CACHE_BASE, 'pkg', hash);
-
-      fs.mkdirSync(tmpFolder, { recursive: true });
-
-      // Example: moduleFolder = /snapshot/appname/node_modules/sharp/build/Release
-      const parts = moduleFolder.split(path.sep);
-      const mIndex = parts.lastIndexOf('node_modules') + 1;
-
-      let newPath;
-
-      // it's a node addon file contained in node_modules folder
-      // we copy the entire module folder in tmp folder
-      if (mIndex > 0) {
-        // Example: modulePackagePath = sharp/build/Release
-        const modulePackagePath = parts.slice(mIndex).join(path.sep);
-        // Example: modulePkgFolder = /snapshot/appname/node_modules/sharp
-        const modulePkgFolder = parts.slice(0, mIndex + 1).join(path.sep);
-
-        // here we copy all files from the snapshot module folder to temporary folder
-        // we keep the module folder structure to prevent issues with modules that are statically
-        // linked using relative paths (Fix #1075)
-        copyFolderRecursiveSync(modulePkgFolder, tmpFolder);
-
-        // Example: /tmp/pkg/<hash>/sharp/build/Release/sharp.node
-        newPath = path.join(tmpFolder, modulePackagePath, moduleBaseName);
-      } else {
-        const tmpModulePath = path.join(tmpFolder, moduleBaseName);
-
-        if (!fs.existsSync(tmpModulePath)) {
-          fs.copyFileSync(modulePath, tmpModulePath);
-        }
-
-        // load the copied file in the temporary folder
-        newPath = tmpModulePath;
-      }
-
-      // replace the path with the new module path
-      args[1] = newPath;
-    }
-
-    return ancestor.dlopen.apply(process, args);
-  };
-})();
+REQUIRE_SHARED.patchDlopen(insideSnapshot);
